requirements.txt
Flask
Pillow
numpy
opencv-contrib-python-headless
gunicorn
